refactoring css, simplyfying
