+++ src/components/ComponentStyles.js
@@ -6,3 +6,10 @@ export const StyledButton = styled.button`
   border-radius: 0.5rem;
   border-color: black;
 `;
+
+export const PageWrapper = styled.div`
+  display: grid;
+  min-height: 100vh;
+  grid-template-rows: 8vh auto;
+  ${(props) => props.theme.backGroundImage()}
+`;
